+++ src/main/java/de/fva_net/maxs/logger/StandardRoutine.java
@@ -1,8 +1,0 @@
-package de.fva_net.maxs.logger;
-
-/**
- * Interface representing a calculation standard routine.
- */
-public interface StandardRoutine {
-    String getId();
-}